Extract users api_token logic to concerns
