+++ app/models/user/api_token.rb
@@ -1,14 +1,8 @@
 # frozen_string_literal: true
 
 class User::ApiToken < ApplicationRecord
-  belongs_to :user
-  has_secure_token
-  validates :user, presence: true
-  validates :name, presence: true, unless: :primary?
-  validates :name, uniqueness: { scope: :user }
-  validates :user, uniqueness: { scope: :primary }, if: :primary?
+  include GobiertoCommon::ActsAsApiToken
 
+  acts_as_api_token_on :user
   delegate :site, to: :user
-
-  scope :primary, -> { where(primary: true) }
 end
